requirements.txt

pymupdf
python-pptx
python-docx